% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCSMC2-package.R
\docType{package}
\name{ABCSMC2-package}
\alias{ABCSMC2}
\alias{ABCSMC2-package}
\title{ABCSMC2: Combined state and parameter inference with approximate Bayesian computation}
\description{
Performs likelihood-free state space inference.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
